@@ -3,46 +3,46 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
-    "@craco/craco": "^5.9.0",
+    "@babel/core": "^7.19.6",
+    "@babel/preset-env": "^7.19.4",
     "@testing-library/jest-dom": "^5.11.4",
-    "@testing-library/react": "^11.1.0",
-    "@testing-library/user-event": "^12.1.10",
-    "@types/jest": "^26.0.15",
-    "@types/node": "^12.0.0",
-    "@types/react": "^16.9.53",
-    "@types/react-datepicker": "^3.1.2",
-    "@types/react-dom": "^16.9.8",
-    "autoprefixer": "9",
-    "firebase": "^8.1.2",
-    "moment": "^2.29.1",
-    "postcss": "7",
-    "react": "^17.0.1",
+    "@testing-library/react": "^13.4.0",
+    "@testing-library/user-event": "^14.4.3",
+    "@types/jest": "^29.2.1",
+    "@types/react": "^18.0.25",
+    "@types/react-datepicker": "^4.8.0",
+    "@types/react-dom": "^18.0.8",
+    "autoprefixer": "^10.4.13",
+    "firebase": "^9.13.0",
+    "luxon": "^3.1.0",
+    "postcss": "^8.4.18",
+    "react": "^18.2.0",
     "react-cool-onclickoutside": "^1.5.8",
-    "react-datepicker": "^3.3.0",
-    "react-dom": "^17.0.1",
+    "react-datepicker": "^4.8.0",
+    "react-dom": "^18.2.0",
     "react-icons": "^4.1.0",
-    "react-router-dom": "^5.2.0",
-    "react-scripts": "4.0.1",
-    "tailwindcss": "npm:@tailwindcss/postcss7-compat",
+    "react-router-dom": "^6.4.3",
+    "react-scripts": "5.0.1",
+    "tailwindcss": "^3.2.1",
     "typescript": "^4.0.3",
-    "web-vitals": "^0.2.4",
-    "workbox-background-sync": "^5.1.3",
-    "workbox-broadcast-update": "^5.1.3",
-    "workbox-cacheable-response": "^5.1.3",
-    "workbox-core": "^5.1.3",
-    "workbox-expiration": "^5.1.3",
-    "workbox-google-analytics": "^5.1.3",
-    "workbox-navigation-preload": "^5.1.3",
-    "workbox-precaching": "^5.1.3",
-    "workbox-range-requests": "^5.1.3",
-    "workbox-routing": "^5.1.3",
-    "workbox-strategies": "^5.1.3",
-    "workbox-streams": "^5.1.3"
+    "web-vitals": "^3.0.4",
+    "workbox-background-sync": "^6.5.4",
+    "workbox-broadcast-update": "^6.5.4",
+    "workbox-cacheable-response": "^6.5.4",
+    "workbox-core": "^6.5.4",
+    "workbox-expiration": "^6.5.4",
+    "workbox-google-analytics": "^6.5.4",
+    "workbox-navigation-preload": "^6.5.4",
+    "workbox-precaching": "^6.5.4",
+    "workbox-range-requests": "^6.5.4",
+    "workbox-routing": "^6.5.4",
+    "workbox-strategies": "^6.5.4",
+    "workbox-streams": "^6.5.4"
   },
   "scripts": {
-    "start": "craco start",
-    "build": "craco build",
-    "test": "craco test",
+    "start": "react-scripts start",
+    "build": "react-scripts build",
+    "test": "react-scripts test",
     "eject": "react-scripts eject"
   },
   "eslintConfig": {
@@ -64,6 +64,8 @@
     ]
   },
   "devDependencies": {
+    "@types/luxon": "^3.0.3",
+    "@types/node": "^18.11.9",
     "@types/react-outside-click-handler": "^1.3.0",
     "@types/react-router-dom": "^5.1.6"
   }
