@@ -1,19 +1,19 @@
 import React from "react";
-import ReactDOM from "react-dom";
+import { createRoot } from "react-dom/client";
 import Provider from "./Context/CategoryContext";
 import "./Styles/styles.css";
-import "react-datepicker/dist/react-datepicker.min.css"
+import "react-datepicker/dist/react-datepicker.min.css";
 
 import App from "./App";
 import * as serviceWorkerRegistration from "./serviceWorkerRegistration";
 
-ReactDOM.render(
+const root = createRoot(document.getElementById("root") as HTMLElement);
+root.render(
   <React.StrictMode>
     <Provider>
       <App />
     </Provider>
-  </React.StrictMode>,
-  document.getElementById("root")
+  </React.StrictMode>
 );
 
 serviceWorkerRegistration.register();
